implement draggablelist and reorder components
