+++ packages/kit/package.json
@@ -5,7 +5,7 @@
   "description": "",
   "exports": {
     ".": "./lib/index.js",
-    "./server": "./lib/server.js"
+    "./server": "./lib/server/index.js"
   },
   "files": [
     "src",
