+++ packages/kit/src/client/Icon.tsx
