+++ packages/kit/src/client/QRCode.tsx
