+++ packages/kit/src/server/Slider.tsx
