+++ packages/kit/src/server/index.ts
@@ -1,16 +1,13 @@
 import { createComponentDefinition } from '@final-ui/react'
-import type RNDraggableFlatList from 'react-native-draggable-flatlist'
-import type { ScaleDecorator as RNSCaleDecorator } from 'react-native-draggable-flatlist'
 import type RNQRCode from 'react-native-qrcode-svg'
 
-import type { Icon } from './Icon'
+import type { RNDraggableFlatList } from '../client/DraggableFlatList'
+import type { Icon } from '../client/Icon'
 
 /** bindings to client-side components */
 export const QRCode = createComponentDefinition<typeof RNQRCode>('RNQRCode')
 export const DraggableFlatList =
   createComponentDefinition<typeof RNDraggableFlatList>('RNDraggableFlatList')
-export const ScaleDecorator = createComponentDefinition<typeof RNSCaleDecorator>('RNSCaleDecorator')
-
 export const RiseIcon = createComponentDefinition<typeof Icon>('RiseIcon')
 
 /** fully server-side components */
